Fixing a couple of Ambiguities

+++ src/ganpdfs/pdformat.py
@@ -138,4 +138,4 @@ class InputPDFs:
                         f, self.x_grid[x], self.q_value
                     )
         reshinpdf = inpdf.reshape((pdf_size, 2 * self.nf + 1, xgrid_size, 1))
-        return reshinpdf
+        return reshinpdf.astype('float32')
